mcp-example-1/requirements.txt
langchain
langchain_google_genai
uvicorn
fastmcp